Updated openstack/openstack Project: openstack/tripleo-heat-templates  d1fcf403dc85b25882c25d36b36ea97b56f05186
Disable systemd-networkd & systemd-resolved

Latest commits in puppet-systemd enabled by default systemd-networkd and
systemd-resolved but we don't want to manage them for now in TripleO.
MySQL and MongoDB services were managing some systemd resources so now
we ensure that these 2 systemd services are disabled. In the future, we
might want and activate these services and revert that patch but for now
we want to disable them.

Change-Id: I42c6c9b643a71a0fbb1768bbae91e8bfa916ea00
Closes-Bug: #1704145