Updated openstack/openstack Project: openstack-dev/cookiecutter  0af64c7c00f3ba0a3af1751c82131fc20c8ed472
Fix errors found from making a project

The pbr version is tied to the repo-name - that is, the name that we
register with PyPI. It takes the argument 'python-novaclient' not
'novaclient'.

Nothing we do is Copyright OpenStack Foundation. There isn't a good
answer for a single-line copyright statement such as is found in sphinx
config files, but "OpenStack Foundation" is definitely wrong.

Add some line wrapping so that the releasenotes conf file doesn't fail
pep8.

Change-Id: Id237331c279b7acb848a28d167ad88d63de6f140