Updated openstack/openstack Project: openstack/sahara-image-elements  83c3011f69389e5e1459e5005134a0ff52e8bfcd
mapr: fix the discovery of the version of Scala

The structure of the page changed slightly between July 8th
and now (thanks Internet Archive).
Adapt the extraction code. including a more robust regex.

Also update the download location (lightbend.com instead of
typesafe.com, even if the latter is still a redirect).

Finally, expose a new variable which allows users to
force a specific version of Scala, disabling the autodetection.
This is useful to workaround possible future similar breakages
without changing the code.

Closes-Bug: #1704337

Change-Id: Ib105d08816d6833f05301e59f64163de5f070109