Updated openstack/openstack Project: openstack/castellan  8980bf7da55dd084ad84c84534fe937f0d43b9c0
Use Stevedore for better extensions

Since all Oslo library drivers are discoverable via
stevedore, we should use stevedore in Castellan as well.
This will make it easier for folks to write their own
custom drivers. Stevedore uses setuptools entry points
for implementing the common patterns for dynamically
loading extensions.

We add [key_manager]/backend as the new option to set
the custom driver. For a while, we should support the
older values that used to be specified using
[key_manager]/apiclass.

Change-Id: I2610459839806a5591da1efa314dfa52bcfb7cda