Updated openstack/openstack Project: openstack/cinder  817f34ce5dfd38d49b909ec1698e1026aee32a18
Pop "consistencygroup" from volume object

Got the following error during the groups tempest test:

DBAPIError exception wrapped from (psycopg2.ProgrammingError)
multiple assignments to same column "consistencygroup_id".

The solution is to pop "consistencygroup" from volume object
before saving to db. This is because consistencygroup is the
name of a relationship in the ORM Volume model, so SQLA tries
to do some kind of update of the foreign key based on the
provided updates if 'consistencygroup' is in updates.

Closes-bug: #1703706
Change-Id: I9489ebd20c0744d69431bc91634e1decffcb9ed9