Updated openstack/openstack Project: openstack/ironic  bc60c21d63b33aa36fc3299e5a91268c47fbeb2a
Generate iPXE boot script on start up

This is a revert of [0], which added some problematic short circuit
logic to prepare_ramdisk. This patch proposes a better fix that was
suggested in the reviews of [0], which is to generate the boot script
on start up rather than during prepare_ramdisk. The iPXE boot script
is not dependent on node properties, or anything that is determined
during deployment, so generating the boot script on start up should
not harm any current functionality.

[0] 2c18f22f9976870b71bfd1e471afa18fd50a7a25

Change-Id: Id545d6cf93227cf1fc2ff0c05dbdceb8fb6aa5b9