Updated openstack/openstack Project: openstack/keystonemiddleware  1c12eb4131d803880a3ab8dda198f6075ec86b3a
Added "warning-is-error" sphinx check for docs

This change adds the "warning-is-error" setting
to setup.cfg in order to enforce strict doc validation which
will cause the build to fail if any warnings are thrown.

This also removes the redundant loading of the todo plugin
warning that shows up while running 'tox -e docs' with
'warning-is-error' enabled.

Change-Id: I33c110073feec7dd38ab75981d6f97c654852f37