Updated openstack/openstack Project: openstack/tripleo-heat-templates  07a5fe1ba66cbad5afafbbd30d2d78129cb90bae
Adds network/cidr mapping into a new service property

Makes it possible to resolve network subnets within a service
template; the data is transported into a new property ServiceData
wired into every service which hopefully is generic enough to
be extended in the future and transport more data.

Data can be consumed in service templates to set config values
which need to know what is the subnet where a deamon operates (for
example the Ceph Public vs Cluster network).

Change-Id: I28e21c46f1ef609517175f7e7ee19e28d1c0cba2