Updated openstack/openstack Project: openstack/nova  2b4c0f54ac826869549e903b37e675dfe9fe0b64
Default reservations=None in Cells v1 and conductor APIs

This is a preliminary change for the counting quotas series to make
any legacy quota "reservations" arguments default to None so that in
the next change, we can simply not pass any reservations instead of
passing empty reservations. Reservations are being deprecated as part
of the counting quotas series.

Part of blueprint cells-count-resources-to-check-quota-in-api

Change-Id: I86e56c1abd979ac9968a3ce7daad07530080bde4