Updated openstack/openstack Project: openstack/neutron  0cd9161c86ed43e6dab04058fd43f89a4e2eca55
Add support for list querying in resource cache

This adjusts get_resources filters to take a tuple of values
instead of a single value. This gets translated into an OR query
on the server side.

This is used in the dependent patch to query for all ports in a
set of security groups.

Change-Id: I10ce263d3eb89cfec3bae4f66d3ef59365a18e15
Partially-Implements: blueprint push-notifications