Updated openstack/openstack Project: openstack/neutron  8ff216d0c1b90e899ff1f5b2e325d13673ddbaa8
DVR: Fix neutron metering agent to notify hosts hosting DVR

DVR router updates are notified based on where they are hosted.
But in the case of metering rpc notifier the notification to a
specific host was not implemented and also when a call is being
made to routers_updated_on_host, it is being ignored since it only
implements the routers_updated function call.

Change-Id: I39b0d43b14294a8eecf6cba230d948c2c45a0b7a
Closes-Bug: #1682345