Updated openstack/openstack Project: openstack/horizon  469bee038d6285cdb9d65187b167b953f46e6ffd
Drop FWaaS related policies from neutron_policy.json

This is the final cleanup related to FWaaS dashboard split out.

Note that 'insert_rule' and 'remove_rule' in neutron_policy.json are
kept as the current policy.json in neutron still has it. They are
specific to neutron-fwaas and theoretically it should be moved from
neutron to neutron-fwaas, but it should happen in neutron side.
As horizon, we keep them until policy.json in the neutron repo has them.
(Related to neutron bug 1703347)

Implement blueprint split-out-neutron-xaas-dashboards

Change-Id: I99b63628365b90f51dc8f76be7b3d63b195916b0