Updated openstack/openstack Project: openstack/diskimage-builder  787e76b9168f1f59497cbca1618aa7060bf5d157
Remove additional Bumblebee repository for opensuse element

The purpose of the openSUSE element is to build openSUSE distribution
based images, so an additional community repo shouldn't be pulled into
the image. In addition the dkms dependency is blacklisted for SUSE
in the dkms element anyway, so this should be a noop.

Change-Id: I0aa06d9f4f110546032f910e3361840693d02de7