Updated openstack/openstack Project: openstack-infra/project-config  73cc4ff92aa75657775d6ac4d24ee52d62e0dc94
neutron-fwaas: Add a non-voting job to run networking-midonet unit tests

It runs networking-midonet unit tests which includes its
neutron-fwaas integration, to expose regressions earlier and
save the effort in both of projects.  Briefly discussed in
a fwaas meeting. [1]

[1] http://eavesdrop.openstack.org/meetings/fwaas/2017/fwaas.2017-06-27-14.02.log.html#l-165

Background: Neutron-fwaas changes sometimes break networking-midonet's
fwaas v1 driver.  It's mainly due to the non-ideal way how the driver
is written.  Ideally we should have a real driver api for fwaas v1.
But the development focus has already moved to fwaas v2.

Note: This job runs the identical set of tests as the voting
counterpart, gate-networking-midonet-python35.  While it includes
tests which are not related to fwaas, it doesn't seem to worth
reducing it to minimum.  We can revisit it if it turns out to be
a problem.

Closes-Bug: #1703328
Change-Id: I0ae979ef0a34f2b75f40f8315cf3f80a0b0c07e0