Updated openstack/openstack Project: openstack/nova  db1406f5ee76d06919c2db8da83cff5d9d3ad6ff
conf: fix netconf, my_ip and host are unclear

Default values set for my_ip and host config opts in the netconf file
are reporting the details of the infra worker, thus making it unclear
what the real default value should be.  Also, help text for the host opt
did not mention its relevance to the cinder and neutron settings.

Change-Id: I69e3953fa46766ea2818bd01c4de949fd43938b0
Closes-Bug: 1669746
Implements: blueprint centralize-config-options-pike