Updated openstack/openstack Project: openstack/nova  3517f6e012b2dccf8763cf330c89626fa35e47a2
fix unshelve notification test instability

The unshelve notification sample test shelve-offloads an instance, waits for
it state to change to SHELVED_OFFLOADED then unshelve the instance and matches
generated the unshelve notification with the stored sample. This test
intermittently fails as the host paramter of the instance doesn't match
sometimes. The reason is that the compute manager during shelve offloading
first sets the state of the instance then later sets the host of the instance.
So the test can start unshelving the instance before the host is cleaned by
the shelve offload code.

The test is updated to not just wait for the state change but also wait for
the change of the host attribute.

Change-Id: I459332de407187724fd2962effb7f3a34751f505
Closes-Bug: #1704423