Updated openstack/openstack Project: openstack/pycadf  fd2a777b03f2ef6dd3753e9f7619adc7f90e4f4e
Turn on warning-is-error for sphinx build

The doc-migration spec requires all projects to treat warnings
as errors during sphinx build. This patches turns that function
on and resolves the error (detailed below) coming as a result
of it.

The code block in event_concept.rst file is currently made to
render in javascript which is an incorrect syntax for it.

Change-Id: Id7fee247b76ed7cfebb849305e0611bc966709be