Updated openstack/openstack Project: openstack/cinder  050a1eda060a44635d67d97013df124848a4ee4b
Migrate configuration-reference to Cinder repo

The next patch in the documentation migration process brings over
the very popular configuration reference guide.  This required
moving the 'block-storage' section from config-reference in
openstack-manuals.  It also brings with it a number of supporting
tables.

With this patch in place it becomes imperative that we make sure
that any changes that touch config options also come with a patch
to this part of Cinder.

As with the other changes, this is intended to just make the
content accessible on the web again.  There are definitely
changes/improvements needed to make this integrate better
with out index page and to ensure the content is accurate.
Later patches will address those needs.

Change-Id: I05c243001e17f18c8b8e9aba91b4f3e2811bf023