Updated openstack/openstack Project: openstack/governance  426b8b7c4f21337b8c241dd6c139e4cc86374589
Draft technical committee vision for public feedback

The following is a vision document that was created by the Technical
Committee. It imagines what the TC and the OpenStack community look
like 2 years out. The goal is to paint a picture of where we want to
be, to help us all know if we're doing things to get us there.

Change-Id: Ifde9f19cfa07c62320e2ce09f6d5e803357aa0b4