Updated openstack/openstack Project: openstack/nova  55825944567416dc802f02a97c77eb4fbf955dc3
Update policy descriptions for base.

These are the final policy rules that are missing a description.

The "network:attach_external_network" policy rule is moved from
base to servers policy since it's not really a base rule, it
applies to creating a server with requested networks or attaching
networking to a server. Therefore it is moved and the documentation
for it are both cleaned up.

Change-Id: I050c4f5f19aa79a682e076cc3e47eba597f272dd
Implements: blueprint policy-docs