Updated openstack/openstack Project: openstack/nova  b40a4be41547627ad2c4a624427ae405cb52d5b1
Remove the test for the route '/resources.:(format)'

Requesting the specific format response should be done by HTTP header
'ACCEPT'. The way of '/resources.:(format)' doesn't support or doc
in the Nova API officially. And all of those kind of URL routes are
removed in the Pike release. This patch removes the related unittest.

Partial-implement-blueprint api-no-more-extensions-pike

Change-Id: I2bda2087ba1f07f6ec0f65713ce75234ff00c0df