Updated openstack/openstack Project: openstack/glance  20ea7ab12098953f9e296b557d6d96afd5e957c9
Satisfy API Reference documentation deleting tags

The API Reference documentation for Glance v2 documents the response
status code for the metadefs delete-tags call as 204, but instead that
call returns a 200 status code. This patch fixes the disparity and adds
the ability to start testing our response serializer.

This patch changes the response code instead of the documentation
because 204 (No Content) is the correct response for this call and
because all the other metadefs DELETE calls return 204.  So to preserve
consistency and to honor the published contract, this patch changes the
code, not the documentation.

This patch was discussed with the API-WG (on 2 Feb 2017) and the TC (on
14 Feb 2017), and they both endorse this approach. See the meeting logs
for details.

Closes-bug: #1656183
Depends-On: I84626976ca729f65b0ee0ea6afe5c9a6a408eecc
Change-Id: I4b6dc1714aeca409a85bfa1f9b729147da704df8