Updated openstack/openstack Project: openstack/heat  17b9aae55813c620dcba6b4149224a3e95dd3655
Cinder volume attachment attributes should not be cached

The Cinder volume 'attachments' and 'attachments_list' attributes are
dynamically built lists and must be queried from the Cinder API on each
request. There are several factors that may change the contents of this
list that are unknown to Heat.  Caching should be prevented for these
attributes.

Closes-Bug: #1703605
Related-Bug: #1703387
Co-Authored-By: Paul Bourke <paul.bourke@oracle.com>
Change-Id: I1a646e567b2d88f2acf09b3077c205c5b061d4bc