Updated openstack/openstack Project: openstack/nova  62f4d6036264febd4a2afa8c652797d5e7a08af7
api-ref: mention disk size limitations in resize flavor

In the resize API reference, mention that the specified flavor
must have a disk size greater than or equal to the current
instance flavor. This is something that doesn't get checked
in the API so it results in a failure on the compute service
which is not obvious to the API user since the instance goes
back into state 'ACTIVE'.

Change-Id: I33fc92ffa346b908a8e4f95da16a68040cfc678e