Updated openstack/openstack Project: openstack/oslo.config  2d0ba09bd685c5dc7c33ddead3ce4fc5a2ae5325
handle option defaults that are not strings

The sample config generator and sphinx integration are breaking when
option defaults are not strings. This shows up in cinder, so this bug
is preventing cinder from adopting the sphinx integration for showing
configuration options.

Fix the rendering in the generator, and in the type class for list
options.

Change-Id: Ib8a248b6dc695b6afe4f1e760af836ac664fa137
Signed-off-by: Doug Hellmann <doug@doughellmann.com>