Updated openstack/openstack Project: openstack/python-ironicclient  035aff25d384c0816176e63ec1c027fbb022f1b4
Update volume release notes to fix reno

The portion of the release note file name with
the hexidecimal value used for tracking was not
unique between the files, and was duplicated from
the API additions.

Renaming should fix the release notes.

Additionally added the API microversion for the volume
interface to the release notes.

Change-Id: If8546ac18b2904396d86547d012d0491f3811a21