Updated openstack/openstack Project: openstack/glance_store  dc650c069bd2c2446400e97e734f18278f1b72ac
Cinder driver: TypeError in _open_cinder_volume

Most of the drivers return a string as the value of device['path']
which is then opened and returned as a file-like object.  When the
volume driver type is 'rbd', the volume of device['path'] is already
a file-like object that can be returned directly.  This patch adds
code to handle this special case.

Change-Id: I0c73a740505420c48a1d1ec7d21449fb8848f6ff
Closes-bug: 1643516