Updated openstack/openstack Project: openstack-infra/project-config  5e65acd573e98093cc0f2ecc7b0d0738adec28c8
validate service types used for documentation jobs

We want the service type use to be consistent across all of the various
repos where we refer to it. This patch adds validation to ensure that
service types associated with jobs for repositories match the service
types assigned by the registry.

In the process, it also changes the job definitions for some of the
projects that were publishing to the "wrong" location. A separate patch
will establish redirects from the old locations to the new locations.

Some of the projects were not listed in the registry at all, so this
patch depends on other patches that add the missing entries.

The validation class defined here will eventually move to the
os-service-types library, when that repository is ready to receive code.

Change-Id: I2785ca8782b592c7af574e0b5a41407610873349
Depends-On: I27765d7760352f4c589168037ddd2e6792d1561f
Depends-On: Ie530edd8aa40e5b89f997526e68117cafa68ee9c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>