Updated openstack/openstack Project: openstack/cinder  08938c9dfdd2d08fda502ed69597ab0039f0b005
Add insufficient space async error in create vol

cinder does not provide an async error message when a new volume
is created from an image if there is insufficient space to download
the image. The changes are made such that if there is an IOError
in fetch() or insufficient space error from check_available_space()
cinder message API generates an async error message for the user.

A new exception ImageTooBig() has been created to handle this
error and whenever this exception is come across, cinder generates
an error message. Changes are made in the create_volume flow to
handle this exception and generate the message.

Changes are made in image_utils.py so that whenever there is not
enough space, the appropriate exception is raised. New action and
detail are added to message_field.py for this type of error.

Change-Id: I184acd9c26d2a2a140c8205e4c93164204e8674d