Updated openstack/openstack Project: openstack/cinder  a920c18557852bbf0960bde5ce7fa49fc4e5569a
Move config-reference to configuration

It was discovered today that the 'config-reference' directory
was supposed to have been renamed to 'configuration'.  The doc
team was expecting it to be at that location so that they could
automatically link to it from contributor/index.html .

This patch just moves the directory.  I am going to be submitting
a subsequent patch that properly sets out the framework for the
additional directories so that we don't have addition problems
in the future.

Change-Id: I6729aaf4593a0b089277f6d7d0aa50fee3ca340f