Updated openstack/openstack Project: openstack/requirements  e7c138ecf22dd227cdb86fbaf9f205aaa01e6511
This change adds os-net-config [0] to global-requirements.txt.

Here are answers to the acceptability questions listed in [1]:

* Is the library actively maintained?

Yes, see for instance [2].

* Is the library good code?

os-net-config follows Openstack codestyle incl. pep8, is used in pro-
duction as part of tripleo installer, and has a high test coverage.

* Is the library python 3 compatible?

Yes.

* Is the library license compatible?

Yes (Apache 2.0).

* Is the library already packaged in the distros we target
(Ubuntu latest / Fedora latest)?

os-net-config is packaged in Debian, Ubuntu and Fedora [3].

* Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

The consuming project is tripleo-validations, which aims to reuse
the schema-based validation of configuration data recently added into
os-net-config to validate the network environment and NIC configuration
files provided for deployment.

* Is the function of this library already covered by other libraries in
global-requirements.txt?

No.

[0] http://git.openstack.org/cgit/openstack/os-net-config/
[1] https://github.com/openstack/requirements#for-new-requirements
[2] https://github.com/openstack/os-net-config/commits/master
[3] https://admin.fedoraproject.org/pkgdb/package/rpms/os-net-config/

Change-Id: I987d1bb15822f3c0d8c72f1ce1559a69cfc41591
Signed-off-by: Frank A. Zdarsky <fzdarsky@redhat.com>