Updated openstack/openstack Project: openstack/neutron  1c94a80b5517f1f536c1c72f112156d4412c6e27
Add callback BEFORE_DELETE for delete_router

Enable a callback to be bound for deletion of routers.
A use case for this is in the NSX plugins, here a load
balancer needs to be connected to an existing router.
If one tries to delete the router then we would like to
prevent that if a load balancer is attached.

Change-Id: Ie0c422578acaa94e8e34c15beca9f18691a3abc6