Updated openstack/openstack Project: openstack/neutron  71d9aab87e37b5162ef09b8cbe3b72709fc88a8b
Only ensure default security group exists if necessary

A default security group will be assigned to a port if
and only if the security group attribute is not set on
the port being created. If the attribute is set then we
do not need all of the extra logic of ensuring that the
default security group exists.

This change save a few cycles, more specifically we do not need to
read from the DB or create a default security group if the
default will not be used.

TrivialFix

Change-Id: Ib072c2753862ce0709a67942527e9c6241130042