Updated openstack/openstack Project: openstack/cinder  f6b6b6550b0022069b741ae0eed4a6bb49ca3bba
Mark Virtuozzo driver as unsupported

The Virtuozzo CI has not reported success past our CI required period.

Checking name: Virtuozzo Storage CI
  first seen: 2017-04-04 08:33:41 (62 days, 12:38:26 old)
  last seen: 2017-04-19 14:23:59 (47 days, 6:48:08 old)
  last success: 2017-04-18 15:30:01 (48 days, 5:42:06 old)
  Job dsvm-tempest-kvm 82% success out of 250 comments S=207, F=43
  Job dsvm-tempest-exe 90% success out of 250 comments S=226, F=24
Overall success rate: 86% of 250 comment

Per Cinder's non-compliance policy, this patch marks the driver
as unsupported and deprecated and it will be removed in the
next cycle if the issue is not corrected.

Change-Id: I1a6a041b7e376c7103bd4413475792d623b48bf9