Updated openstack/openstack Project: openstack-infra/project-config  34a7c5718f3214c17d2e279f2d5a93cdbf7995ab
Add os_tacker role to OpenStack-Ansible

This patch adds a preseeded repository to OpenStack-Ansible for
deploying Tacker.

This project will have its own core team with the general
openstack-ansible core team included, with the appropriate rights to
ensure that they can act as a back-stop in case of any issues.

Change-Id: I111d829c0a4ee698d5f927972f1d3c5e6cda5819
Needed-By: I6e594309b3e62597e08f56dfe793b3eaa6c2598a