Updated openstack/openstack Project: openstack/openstack-manuals  a885ce5c900f93448bd1ab38c34c7fd119078218
be consistent with how we list libraries

The library projects tend to have longer descriptions and no real
service type. This makes them look odd when listed in the admin guides
or configuration guides pages. On the openstack-projects.html page we
list libraries by their name (as a link) with a description below. Apply
that same style formatting to the admin and config guide lists for
consistency.

Change-Id: Idfb53469a9df5c8d05a3f85a92e9ad2dec3185a0
Signed-off-by: Doug Hellmann <doug@doughellmann.com>