Updated openstack/openstack Project: openstack/horizon  299f4c3562593e8e99e04fab2a539032adcdb8d9
Fixes to allow us to use Django 1.11

Mostly the issue with Horizon was we were using some
undocumented widgets that were cut for version 1.11
django. This patch adds their code into our repo for
now, but we need to do a proper rewrite of our widgets
for Queens to use current widgets in Django.

Also edits some tests because the functionality changed
in Django 1.11

Implements: blueprint dj111
Co-Authored-By: Rob Cresswell <robert.cresswell@outlook.com>
Change-Id: I444d45f274662f4f33701c16cce4ae80cb546654