Updated openstack/openstack Project: openstack/requirements  b29c2d152c7535576cebc820cc612d1adb079823
Remove docker-py from global requirements

The pypi package 'docker-py' [1] has been renamed to 'docker' [2].
It is better to move to the new 'docker' package because the old
package will be deprecated and all the new features will go into
the new package only.

Package 'docker' has been added to requirements [3]. The old
package 'docker-py' is still allowed to be in the global requirements
during the transition period but it should be removed after all or
most of the projects finish the migration.

Allow both 'docker' and 'docker-py' in the requirement is proven
to be problematic because these two packages are not compatible
with each other (see bug #1693425). Ideally, we should be able
to drop 'docker-py' as soon as possible

[1] https://pypi.python.org/pypi/docker-py
[2] https://pypi.python.org/pypi/docker
[3] https://review.openstack.org/#/c/423715/

Change-Id: I8b695b7fa9dfcc991688ffcfe850c347a3cbd0d9
Depends-On: I3d952f647f8235079f717256750c90cb9ddf9588
Depends-On: I9346112d8a495bc6b304647377f8da47017cc869
Depends-On: I0a6fdb905016ff064d90b3170c3bcf90311aaaad
Depends-On: I72951c2d3a19b0bcfc185d0cd7b9e30ba5d40b25
Depends-On: Id440e7362a65ddd56cfb77f3a8b502750c283a21
Depends-On: I50bea360878ffb2450f08627243b139c241b33e8
Depends-On: I3ffd8bb6461e8b7edacc5e833754f00c4249b854
Depends-On: Ibcd5a57a1fbf55dcc5a690e41f20917f95b63da0
Related-Bug: #1693425