Updated openstack/openstack Project: openstack/nova  2e0a90601414dd0513bebdcf6bc25d3334e16410
XenAPI: Fix ValueError in test_slave_asks_master_to_add_slave_to_pool

Change If9e8dd5cc2634168910d5f9f8d9302aeefa16097 in
oslo.serialization will raise a ValueError when it can't
convert something to a primitive, which in this case is the
Aggregate object used in this test. Given the add_to_aggregate
method doesn't even serialize the Aggregate object passed in,
I'm not sure how this was working before, but the fix is to
just not try and expect add_aggregate_host RPC API is called
with a serialized Aggregate object.

Change-Id: I139533dfd94a1aa7a00a6220a4411bebe1c9e88b
Related-Bug: #1593641