Updated openstack/openstack Project: openstack/nova  7d77eb465869382bb265f4731a99527f6b774cc5
Convert HostState.limits['numa_topology'] to primitive

The SchedulerManager.select_destinations method calls
_host_state_obj_to_dict to convert a list of selected
HostState objects to dicts before converting to primitives
and sending over RPC.

The problem is the NUMATopologyFilter can set the
HostState.limits['numa_topology'] field to an instance of
a NUMATopologyLimits versioned object. That was not being
converted to a primitive before being sent back over RPC.

The jsonutils.to_primitive() method will simply return an
object that it does not know how to convert to a primitive.
As of change If9e8dd5cc2634168910d5f9f8d9302aeefa16097, however,
to_primitive() will now raise a ValueError. This was failing
in nova functional tests in I76a90a57070c5184e893a79406b5d1784fcc969f.

We need to convert the NUMATopologyLimits to a primitive before
calling jsonutils.to_primitive() to avoid the ValueError. This
change explicitly only cares about this one known value and does
not attempt to do this generically for out of tree filters or
scheduler drivers.

Change-Id: I01e4cc2ca9d193b64023bc7df04ef878e68d46b8
Closes-Bug: #1593641