Updated openstack/openstack Project: openstack/python-novaclient  8d6bde276b6c930fe75320f75706e180bfab4495
Add resize down test which also verifies quota changes

This adds a functional test which creates a server at a larger
flavor and then resizes it down to a smaller flavor, and then
reverts the resize. It checks quota before, during and after
the resize. This is an important wrinkle with the counting
quotas changes happening in the server in Pike.

Change-Id: If1bb59e513b91a99fae2b6825248a40d8e7ac7b9